debug-detective/requirements.txt
bedrock-agentcore==1.4.6
bedrock-agentcore-starter-toolkit==0.3.3
strands-agents==1.30.0
strands-agents-tools==0.2.22
